+++ packages/alfa-css/test/value/shape.spec.ts
@@ -1,23 +1,14 @@
-import { Assertions, test } from "@siteimprove/alfa-test";
-
-import { Slice } from "@siteimprove/alfa-slice";
+import { test } from "@siteimprove/alfa-test";
 
 import { Lexer } from "../../src/syntax/lexer";
 import { Shape } from "../../src/value/shape";
 
-function parse(t: Assertions, input: string, expected: Shape.JSON) {
-  t.deepEqual(
-    Shape.parse(Slice.of(Lexer.lex(input)))
-      .map(([_, shape]) => shape)
-      .get()
-      .toJSON(),
-    expected,
-    input
-  );
+function parse(input: string) {
+  return Shape.parse(Lexer.lex(input)).map(([_, shape]) => shape.toJSON());
 }
 
-test("parse() parses <basic-shape> <geometry-box>", (t) => {
-  parse(t, "inset(1px) padding-box", {
+test(".parse() parses <basic-shape> <geometry-box>", (t) => {
+  t.deepEqual(parse("inset(1px) padding-box").get(), {
     type: "shape",
     box: {
       type: "keyword",
@@ -39,8 +30,8 @@ test("parse() parses <basic-shape> <geometry-box>", (t) => {
   });
 });
 
-test("parse() parses <geometry-box> <basic-shape>", (t) => {
-  parse(t, "margin-box inset(1px)", {
+test(".parse() parses <geometry-box> <basic-shape>", (t) => {
+  t.deepEqual(parse("margin-box inset(1px)").get(), {
     type: "shape",
     box: {
       type: "keyword",
@@ -62,8 +53,8 @@ test("parse() parses <geometry-box> <basic-shape>", (t) => {
   });
 });
 
-test("parse() parses <basic-shape>", (t) => {
-  parse(t, "inset(1px)", {
+test(".parse() parses <basic-shape>", (t) => {
+  t.deepEqual(parse("inset(1px)").get(), {
     type: "shape",
     box: {
       type: "keyword",
@@ -85,6 +76,6 @@ test("parse() parses <basic-shape>", (t) => {
   });
 });
 
-test("parse() fails if no <basic-shape> is provided", (t) => {
-  t.deepEqual(Shape.parse(Slice.of(Lexer.lex("margin-box"))).isErr(), true);
+test(".parse() fails if no <basic-shape> is provided", (t) => {
+  t.deepEqual(parse("margin-box").isErr(), true);
 });